sentiment_analysis_agent/requirements.txt
google-adk==1.2.1
google-cloud-aiplatform[agent_engines,adk]